+++ 13-suspense/04-ux/.eslintrc.js
@@ -4,27 +4,23 @@ module.exports = {
     es2020: true,
   },
   extends: [
-    'plugin:react/recommended',
     'airbnb',
     'airbnb/hooks',
     'plugin:import/errors',
     'plugin:import/warnings',
     'plugin:import/typescript',
-    'plugin:@typescript-eslint/eslint-recommended',
     'plugin:@typescript-eslint/recommended',
     'plugin:@typescript-eslint/recommended-requiring-type-checking',
-    'plugin:prettier/recommended',
     'prettier',
     'prettier/@typescript-eslint',
     'prettier/react',
-    'prettier/standard',
   ],
   parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaFeatures: {
       jsx: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 12,
     project: './tsconfig.eslint.json',
     sourceType: 'module',
     tsconfigRootDir: __dirname,
@@ -34,12 +30,16 @@ module.exports = {
     'import',
     'jsx-a11y',
     'prefer-arrow',
-    'prettier',
     'react',
     'react-hooks',
   ],
   root: true,
   rules: {
+    // occur error in `import React from 'react'` with react-scripts 4.0.1
+    'no-use-before-define': 'off',
+    '@typescript-eslint/no-use-before-define': [
+      'error',
+    ],
     'lines-between-class-members': [
       'error',
       'always',
@@ -47,10 +47,12 @@ module.exports = {
         exceptAfterSingleLine: true,
       },
     ],
-    // should be rewritten as `['error', { allowAsStatement: true }]` in ESLint 7 or later
-    // SEE: https://github.com/typescript-eslint/typescript-eslint/issues/1184
-    'no-void': 'off',
-    'no-use-before-define': 'off',
+    'no-void': [
+      'error',
+      {
+        allowAsStatement: true,
+      },
+    ],
     'padding-line-between-statements': [
       'error',
       {
@@ -101,6 +103,7 @@ module.exports = {
         explicitSpread: 'ignore',
       },
     ],
+    'react/react-in-jsx-scope': 'off',
   },
   overrides: [
     {