+++ 13-suspense/04-ux/src/containers/oraganisms/MemberList.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import { FC } from 'react';
 import { useQuery } from 'react-query';
 
 import MemberList from 'components/organisms/MemberList';
@@ -6,9 +6,13 @@ import { getMembers } from 'domains/github';
 // import getMembers from 'domains/github/services/get-members-delayed';
 
 const EnhancedMemberList: FC<{ orgCode: string }> = ({ orgCode }) => {
-  const { data: users = [] } = useQuery([orgCode, 'members'], getMembers, {
-    enabled: orgCode.length >= 2,
-  });
+  const { data: users = [] } = useQuery(
+    [orgCode, 'members'],
+    () => getMembers(orgCode),
+    {
+      enabled: orgCode.length >= 2,
+    },
+  );
 
   return <MemberList users={users} />;
 };