+++ 13-suspense/04-ux/src/containers/oraganisms/OrgInfo.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import { FC } from 'react';
 import { useQuery } from 'react-query';
 
 import OrgInfo from 'components/organisms/OrgInfo';
@@ -7,7 +7,7 @@ import { getOrganization } from 'domains/github';
 const EnhancedOrgInfo: FC<{ orgCode: string }> = ({ orgCode }) => {
   const { data: org = null } = useQuery(
     [orgCode, 'organization'],
-    getOrganization,
+    () => getOrganization(orgCode),
     { enabled: orgCode.length >= 2 },
   );